Fix enum field default constant path
The cpp EnumValueDesciptor->full_name() value
is beholden to the c++ oddity where the enum
type is transparent within the enclosing 
message type or namespace. In the ruby implementation,
this is not valid and so we need to actually
insert the enum type's name into the namespace
path to the actual enum default value constant.

Rework message and enum field generation
to re-open the ruby class body and apply the DSL
methods in the class context, rather than calling
from the fully qualified class namespace itself.
This change makes the generated files much easier
to read and more rubyish.