Simplify Message#to_hash, #inspect implementations
The hash representation of a message now contains
only values that have been set previously. Instead
of attempting coercion in the message implementation
we delegate out to the value to make the conversions,
if any.

Add support to FieldArray for Message#to_hash.

Clean up and add simpler specs to Message#to_hash.