Refactor Service API, add ServiceDispatcher
Add ServiceDispatcher class to take care of the core concern of sending
calls to a service method when a server request has been received. The
ServiceDispatcher is also responsible for retrieving the response from
the service and ensuring the response type is appropriate.

Previously, the ServiceDispatcher functionality was shared between the
Server and the Service. The new role of the Server is to simply inflate
the request into the wrapper from the socket and hand it to the
ServiceDispatcher. By invoking the dispatcher, the server then
serializes back to the socket either the response or error.

The new role of the Service is to simply act like a PORO. The rpc DSL
is still defined on the Service but all method calling and
request/response manipulation is left up to the dispatcher. This
significantly simplifies both server and service code and allows us to
move away from the hook implementaiton used before using
`method_added`, aliasing methods and hiding their implementations.

One feature that has been removed from the server side due to lack of
use in practical applications is async responding. While a "nice to
have", it significantly complicated the process of crafting responses
to rpc requests. Async responding may be back in a future version, but
I would imagine we'd use something closer to an actor implementation or
a Fiber-based approach.

All in all, this feels like a much cleaner approach.