Remove client async capabilities
The async client calls are the perfect example of a cool idea built
before a necessary implementation was needed. I'm removing it for now
from the client API in favor of cleaner, easier to test (and refactor)
code.

I imagine this could make it's way back into this gem in the form of
celluloid actors or similar.