Increase directory lookup speed
This improves lookup speed in a number of ways. The mutex locking has
been removed, which I believe is the primary culprit of the increased
lookup times. Writing new listings only happens in a single thread, so
the mutex was an unnecessary precaution.

Additionally, the listings are now hashed by service name for the lookup
method, removing all iteration from the lookup process (except perhaps
on a hash collision).