First cut at an Enum API refactor
== Aliases

Extend Optionable module to get class-level option management. This
allows us to support the allow_alias protobuf option in libprotoc
v2.5.0.

Adds methods to return an array of EnumValue objects corresponding to a
given number "value".

== "Value"

The term "value" is used to mean multiple things in the Enum
implementation: 1) an EnumValue object, and 2) an integer commonly
called a tag number.

This cut attempts to use the term "number" for arguments when the arg
should be number-like, and use the term `enum_value` in methods and
comments when we are referencing the actual object type. All method
name changes are aliased to their old names.

Ideally EnumValue would not be separate from Enum. I'm not willing to
make that cut here, possibly in a future API refactor cut.