Add protobuf:compile rake task for external use
The `protobuf:compile` task can be utilized in any project or gem that
uses protobuf. Open up the Rakefile of the including project and add:

```ruby
require ‘protobuf/tasks’
```

The `compile` task accepts three arguments: package, source, and
destination. Package does not have a default value and must be
specified. Source’s default is ‘defintitions’. Destination’s default is
`lib`. Plugin’s default is `ruby`, and is overridable in case you want
to utilize the descriptors to make your own plugin generator.

The composed command to protoc is printed to the terminal and then the
process is exec’d with that command.