Consolidate response encoding
Instead of returning a response object and requiring each server to
encode it, encode the response in handle_client and return it. That
way, each server is simply responsible for streaming the data as it sees
fit.

It also allows stats tracking and logging to be centralized inside the
server module.