Use sane names in Enum class
Get rid of all references to the word “value”. EnumValue’s are just
called enums now. “value” objects when referring to an integer number
are just now called tags. It’s not semantically correct, but there
isn’t a semantically correct name and “tag” is short. Replace “by” in
method names with “for” cause it reads nicer.