Message.to_json returns the name of the class
This fixes a long-standing bug caused by circular references in the way
the Field system is implemented.

The spec wraps the call to to_json in a timeout because the recursive
checks are unbounded and will eat memory and cpu forever. The simplest
solution is to simply return the name of the message when asked for the
json representation at the class level.