Only rescue specific no method errors in the dispatcher
In the service dispatcher, we're rescuing NoMethodError and returning an
RPC MethodNotFound error assuming that the method has been defined, but
not implmented. This is the correct behavior, but the implementation
catches all NoMethodError exceptions that are raised, which masks real
issues unrelated to the RPC method not being implemented.

Instead, only re-wrap NoMethodError exceptions if the message contains
the RPC method. Raise the original error if it doesn't.