Fix transmission and shutdown
Switched to a dealer socket in the workers so they can send multiple
replies without waiting for a response. Because new work is coming in
through the queue, they no longer need to receive data from the broker
on the zmq socket.

The server was hanging on shutdown. I'm not sure why, but I suspect
reading from the queue was blocking it. I updated it to use the queue
itself to signal shutdown. The worker will also propogate the shutdown
signal when it receives it, so the signal will continue to propogate
until all workers shutdown. That's the theory at least.