Remove memoization on encode
Changing values in repeated fields does not clear the memoized encoded
value. This means that changes to repated fields after the memoized
encoding has been computed will never be encoded. The pattern works for
other field types because the memoized value is being cleared in the
setter on the message before we call into the field. The field itself is
not involved in clearing the memoized value at all.

Repeated fields are a different story. If we were completely resetting a
repeated field to a new array, the memoized encoding would be correctly
cleared, but because the field array we are pushing into doesn't have any
concept of a message, let alone the specific message instance the field
belongs to, there is no way to clear the memoized encoded message.

We need to remove the memoization until we can find a better approach.

Fixes #304.