Field values should be stored via their fully qualified names
This fixes a bug where ::Protobuf::Message#{simple_field_name}! would return nil rather than the value of the field, if the field was an extension.
This also fixes a bug where ::Protobuf::Message.field?(name) would always return false if name was the fully qualified name of an extension.