Log error in run loop before teardown to pevent freezing without error
It's possible for the teardown call to fire and wait on workers to
finish (which might never finish). In this case, we're never going to
see this error. To combat this, we can log the error before the ensure
block is executed. The fact that we found this code path means that
someone is not handing an error further down. But until we know which
error we're missing, we can't fix it.