Update the compile task to work with protoc v3
The Protocol Buffers v3 compiler (protoc) now comes with a builtin Ruby
plugin. This conflicts with our local custom Ruby plugin. In order to
use it, the custom plugin must be passed to the compiler, and invoked
via the custom output generator:

  protoc
    --plugin=protoc-gen-ruby-protobuf=/path/to/bin/protoc-gen-ruby
    --ruby-protobuf_out=lib

Update the compile task to take this new pattern into account so that it
works with both the v2 and v3 compilers.